LayoutTests/inspector/styles/protocol-css-regions-commands-expected.txt

Tests the following commands:

getNamedFlowCollection Bug 91607
getFlowByName Bug 91855


Running: testGetNamedFlowCollection
=== CSS Named Flows ===
flow1
flow2
flow3

Running: testGetFlowByName1
=== Named Flow "flow2" from main document ===
name: flow2
overset: false

Running: testGetFlowByName2
=== Name Flow "flow4" from main document ===
There is no Named Flow "flow4" in the main document
